requirements.txt
Flask-APScheduler==1.13.1
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.11
python-dotenv==1.2.2
Flask-Session==0.8.0
SQLAlchemy==2.0.48
reportlab==4.4.10
requests==2.33.1
asgiref==3.11.1
urllib3==2.6.3
bcrypt==5.0.0
Flask==3.1.3
pyotp==2.9.0
qrcode==8.2
